feat(1091): change templateId to integer in migration files

+++ migrations/20200204-add-column_templateId_to_builds.js
@@ -9,7 +9,7 @@ module.exports = {
     up: async (queryInterface, Sequelize) => {
         await queryInterface.sequelize.transaction(async (transaction) => {
             await queryInterface.addColumn(table, 'templateId', {
-                type: Sequelize.DOUBLE }, { transaction });
+                type: Sequelize.INTEGER }, { transaction });
 
             await queryInterface.addIndex(table, 'templateId', {
                 name: `${table}_template_id` }, { transaction });
